ActivityLog: splitted ActivityLogMessageParser into several files per each module
