+++ webapp/loggedin/modules/home/dashboard/activityLog/activityLogMessageParser/activityLogMessageParserUtils.js
@@ -1,0 +1,9 @@
+import * as R from 'ramda'
+
+import * as ActivityLog from '@common/activityLog/activityLog'
+import * as Survey from '@core/survey/survey'
+
+export const getNodeDef = survey => R.pipe(
+  ActivityLog.getContentUuid,
+  nodeDefUuid => Survey.getNodeDefByUuid(nodeDefUuid)(survey)
+)
