+++ webapp/loggedin/modules/home/dashboard/activityLog/activityLogMessageParser/isItemDeletedFns/isItemDeletedFnsSurvey.js
@@ -1,0 +1,17 @@
+import * as R from 'ramda'
+
+import * as ActivityLog from '@common/activityLog/activityLog'
+
+import * as ActivityLogMessageParserUtils from '../activityLogMessageParserUtils'
+
+export default {
+  [ActivityLog.type.nodeDefCreate]: survey => R.pipe(
+    ActivityLogMessageParserUtils.getNodeDef(survey),
+    R.isNil
+  ),
+
+  [ActivityLog.type.nodeDefUpdate]: survey => R.pipe(
+    ActivityLogMessageParserUtils.getNodeDef(survey),
+    R.isNil
+  ),
+}
